celery_worker/requirements.txt
# ============================================
# Python Celery Worker Dependencies
# Save as: celery_worker/requirements.txt
# ============================================

# Celery and dependencies
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Database
pymongo==4.6.1

# Image processing
Pillow==10.1.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Email and SMS
twilio==8.11.1

# Cloud storage
cloudinary==1.37.0

# Monitoring (optional but recommended)
flower==2.0.1

# Utilities
python-dateutil==2.8.2